requirements.txt
fastapi[standard]
dotenv
google-genai
fastembed
pinecone
rank-bm25